Moved unit tests to external module, dropped changes to degrees symbol formatting
